fix(chat_agent): fix a bug on prod that saves/fetches data from db as str instead of dict
this issue is weird to me since it's only happening on prod. the current workaround is to load the data as json if it's not of type 'dict'. it could be how the proxy forwards the data to gunicorn, or how flask interprets the data, or how the SQLMessageHistory saves the data into the db.