feat(generate): change the behaviors of save_classes_and_properties() and get_class_by_name() functions
these changes were made due to a bug where an existing class won't be saved to db and hence will disregard the entire props and instances from being saved to db. the behavior after these changes will be to check if a class (with the same name) under the same conversation has already been made and if so the system will retrieve the class from db and only fetch the class_id without creating and saving new instance to db (this behavior should better follow the original design in the ER-diagram).